requirements.txt
gensim
h5py
numpy
torch==0.4.0
pyyaml